clean up obs space creation logic

+++ omnigibson/objects/controllable_object.py
@@ -281,9 +281,7 @@ class ControllableObject(BaseObject):
             low.append(np.array([-np.inf] * controller.command_dim) if limits is None else limits[0])
             high.append(np.array([np.inf] * controller.command_dim) if limits is None else limits[1])
 
-        return gym.spaces.Box(
-            shape=(self.action_dim,), low=np.concatenate(low), high=np.concatenate(high), dtype=np.float32
-        )
+        return gym.spaces.Box(shape=(self.action_dim,), low=np.concatenate(low), high=np.concatenate(high), dtype=float)
 
     def apply_action(self, action):
         """
